+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.3 (unreleased)
 
 - Updated Vega-Embed to 6.20.1 to fix spec version warnings
+- Fixed error with iRuby
 
 ## 0.2.2 (2021-11-06)